+++ CHANGELOG.md
@@ -9,7 +9,8 @@
   Scalyr API and that the API key is valid. This ensures that the plugin doesn't start and
   start consuming events until we can successfully perform a connectivity check which means
   we can't end up in situation when we could potentially drop some events in case of an invalid
-  API token or similar when reaching retry limit and DLQ disabled.
+  API token or similar when reaching retry limit and DLQ disabled. If you want to disable this
+  check on register, you can set ``perform_connectivity_check`` config option to ``false``.
 
 ## 0.2.0.beta